feat: [#651] bumped _adaptix, added documentation
