+++ docs/examples/dealing_with_type_checking/message.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from .chat import Chat
+
+
+@dataclass
+class Message:
+    id: int
+    text: str
+    chat: "Chat"
